refactor: Use db model index type for analytics column (#16346)
