+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/table/model/AnalyticsIndex.java
@@ -29,6 +29,7 @@ package org.hisp.dhis.analytics.table.model;
 
 import java.util.ArrayList;
 import java.util.List;
+import org.hisp.dhis.db.model.IndexType;
 
 /**
  * Class representing an index on a database table column.
